fix: split browser port availability tests for lint line limit
